myblog/requirements.txt

Django==4.2.21
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers



